feat!: Configure backend for Postgres
Created a recipe application and configured the app to use Postgres database. Also created a index route that returns a JSON response to show that the server is running

BREAKING CHANGE: Database login credentials must be created in .env file and a Docker image, or some other instance of Postgres must be connected for the app to start.